Move damp, c2d, rlocus from matlab into control
