+++ control/rlocus.py
@@ -54,6 +54,8 @@ from . import xferfcn
 from .exception import ControlMIMONotImplemented
 from functools import partial
 
+__all__ = ['root_locus', 'rlocus']
+
 # Main function: compute a root locus diagram
 def root_locus(sys, kvect=None, xlim=None, ylim=None, plotstr='-', Plot=True,
                PrintGain=True):
@@ -208,3 +210,5 @@ def _RLFeedbackClicks(event, sys):
     if abs(K.real) > 1e-8 and abs(K.imag/K.real) < 0.04:
         print("Clicked at %10.4g%+10.4gj gain %10.4g damp %10.4g" %
               (s.real, s.imag, K.real, -1*s.real/abs(s)))
+
+rlocus = root_locus
